split out some stuff
